# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BarberPro API",
    "version": "1.0.0",
    "description": "Documentação da API pública do BarberPro. Endpoints para serviços, barbeiros, disponibilidade e agendamentos."
  },
  "servers": [
    { "url": "/", "description": "Servidor atual" }
  ],
  "tags": [
    { "name": "Services", "description": "Catálogo de serviços" },
    { "name": "Barbers", "description": "Listagem de barbeiros" },
    { "name": "Availability", "description": "Consulta de horários" },
    { "name": "Appointments", "description": "Criação de agendamentos" }
  ],
  "paths": {
    "/api/services": {
      "get": {
        "tags": ["Services"],
        "summary": "Listar serviços ativos",
        "responses": {
          "200": {
            "description": "Lista de serviços",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "services": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/ServicePreview" }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Services"],
        "summary": "Criar serviço",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number", "format": "float" },
                  "duration": { "type": "integer", "description": "minutos" },
                  "category": { "type": "string" },
                  "isActive": { "type": "boolean" }
                },
                "required": ["name", "price", "duration"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Serviço criado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "service": { "$ref": "#/components/schemas/ServicePreview" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/services/{id}": {
      "get": {
        "tags": ["Services"],
        "summary": "Buscar serviço por ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Serviço encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "service": { "$ref": "#/components/schemas/ServicePreview" }
                  }
                }
              }
            }
          },
          "404": { "description": "Serviço não encontrado" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "put": {
        "tags": ["Services"],
        "summary": "Atualizar serviço",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Campos a serem atualizados; todos opcionais",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number", "format": "float" },
                  "duration": { "type": "integer", "description": "minutos" },
                  "category": { "type": "string" },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Serviço atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "service": { "$ref": "#/components/schemas/ServicePreview" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "description": "Serviço não encontrado" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "tags": ["Services"],
        "summary": "Excluir ou desativar serviço",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Serviço deletado ou desativado" },
          "404": { "description": "Serviço não encontrado" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/barbers": {
      "get": {
        "tags": ["Barbers"],
        "summary": "Listar barbeiros ativos",
        "responses": {
          "200": {
            "description": "Lista de barbeiros",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "barbers": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Barber" }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "tags": ["Barbers"],
        "summary": "Criar barbeiro",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" },
                  "phone": { "type": "string" }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Barbeiro criado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "barber": { "$ref": "#/components/schemas/Barber" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/barbers/{id}": {
      "get": {
        "tags": ["Barbers"],
        "summary": "Buscar barbeiro por ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Barbeiro encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "barber": { "$ref": "#/components/schemas/Barber" }
                  }
                }
              }
            }
          },
          "404": { "description": "Barbeiro não encontrado" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "put": {
        "tags": ["Barbers"],
        "summary": "Atualizar barbeiro",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Campos a serem atualizados; todos opcionais",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" },
                  "phone": { "type": "string" },
                  "isActive": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Barbeiro atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "barber": { "$ref": "#/components/schemas/Barber" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "description": "Barbeiro não encontrado" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "tags": ["Barbers"],
        "summary": "Desativar ou excluir barbeiro",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Barbeiro desativado ou deletado" },
          "404": { "description": "Barbeiro não encontrado" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/availability": {
      "get": {
        "tags": ["Availability"],
        "summary": "Consultar horários disponíveis por barbeiro e data",
        "parameters": [
          {
            "name": "barberId",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "format": "uuid" },
            "description": "ID do barbeiro"
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": { "type": "string", "example": "2025-02-01" },
            "description": "Data (YYYY-MM-DD)"
          },
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 30 },
            "description": "Duração desejada em minutos"
          }
        ],
        "responses": {
          "200": {
            "description": "Horários disponíveis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "availableTimes": {
                      "type": "array",
                      "items": { "type": "string", "example": "14:30" }
                    }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/appointments": {
      "post": {
        "tags": ["Appointments"],
        "summary": "Criar um agendamento",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AppointmentCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agendamento criado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "appointment": { "$ref": "#/components/schemas/AppointmentSummary" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ServicePreview": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "price": { "type": "number", "format": "float" },
          "duration": { "type": "integer", "description": "minutos" },
          "description": { "type": "string" }
        },
        "required": ["id", "name", "price", "duration"]
      },
      "Barber": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        },
        "required": ["id", "name", "email"]
      },
      "AppointmentSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "date": { "type": "string", "example": "2025-02-01" },
          "status": { "type": "string", "enum": ["pending", "confirmed", "completed", "cancelled"] },
          "client": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "email": { "type": "string", "format": "email" }
            },
            "required": ["name", "email"]
          },
          "barber": {
            "type": "object",
            "properties": { "name": { "type": "string" } },
            "required": ["name"]
          },
          "service": {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "price": { "type": "number" },
              "duration": { "type": "integer" }
            },
            "required": ["name", "price", "duration"]
          }
        },
        "required": ["id", "date", "status", "client", "barber", "service"]
      },
      "AppointmentCreateRequest": {
        "type": "object",
        "description": "Envie 'date' + 'time' OU um único 'dateTime'.",
        "properties": {
          "clientName": { "type": "string" },
          "clientEmail": { "type": "string", "format": "email" },
          "clientPhone": { "type": "string" },
          "clientWhatsapp": { "type": "string" },
          "serviceId": { "type": "string", "format": "uuid" },
          "barberId": { "type": "string", "format": "uuid" },
          "date": { "type": "string", "example": "2025-02-01" },
          "time": { "type": "string", "example": "14:30" },
          "dateTime": { "type": "string", "format": "date-time" },
          "paymentMethod": { "type": "string", "example": "Dinheiro" },
          "payOnline": { "type": "boolean" },
          "notes": { "type": "string" }
        },
        "required": ["clientName", "clientEmail", "clientPhone", "serviceId", "barberId"]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Parâmetros inválidos ou ausentes",
        "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string" } }, "required": ["error"] } } }
      },
      "ServerError": {
        "description": "Erro interno do servidor",
        "content": { "application/json": { "schema": { "type": "object", "properties": { "error": { "type": "string" } }, "required": ["error"] } } }
      }
    }
  }
}